requirements.txt
# Core Streamlit and UI
streamlit==1.38.0
Pillow==10.0.0

# Environment and Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# LangChain Core
langchain==0.2.16
langchain-core==0.2.16
langchain-community==0.2.16

# LangChain Integrations
langchain-openai==0.1.22
langchain-ollama==0.1.3
langchain-anthropic==0.1.23

# Vector Database and Embeddings
faiss-cpu==1.7.4
chromadb==0.5.3

# Document Processing
pymupdf==1.24.10
pypdf==4.3.1
mammoth==1.8.0
docx2txt==0.8

# Data Processing
pandas==2.2.2
numpy==1.26.4

# LLM and AI
openai==1.42.0
ollama==0.3.1
tiktoken==0.7.0

# Visualization (for analysis module)
matplotlib==3.9.0
seaborn==0.13.2

# Utilities
tqdm==4.66.1